+++ controllers/suite_int_test.go
@@ -101,7 +101,9 @@ var _ = BeforeSuite(func() {
 
 	dependencyValidator := dependency.NewCompositeDependencyValidator(&version, &EtcdDoguRegistry)
 	serviceAccountCreator := &mocks.ServiceAccountCreator{}
-	serviceAccountCreator.Mock.On("CreateServiceAccounts", mock.Anything, mock.Anything, mock.Anything).Return(nil)
+	serviceAccountCreator.Mock.On("CreateAll", mock.Anything, mock.Anything, mock.Anything).Return(nil)
+	serviceAccountRemover := &mocks.ServiceAccountRemover{}
+	serviceAccountRemover.Mock.On("RemoveAll", mock.Anything, mock.Anything, mock.Anything).Return(nil)
 
 	doguSecretHandler := &mocks.DoguSecretsHandler{}
 	doguSecretHandler.On("WriteDoguSecretsToRegistry", mock.Anything, mock.Anything).Return(nil)
@@ -116,6 +118,7 @@ var _ = BeforeSuite(func() {
 		DoguRegistrator:       doguRegistrator,
 		DependencyValidator:   dependencyValidator,
 		ServiceAccountCreator: serviceAccountCreator,
+		ServiceAccountRemover: serviceAccountRemover,
 		DoguSecretHandler:     doguSecretHandler,
 	}